refactor: deps status
