+++ deps/status/src/checkDepsStatus.ts
