+++ exec/plugin-commands-script-runners/src/run.ts
@@ -10,6 +10,7 @@ import { prepareExecutionEnv } from '@pnpm/plugin-commands-env'
 import { FILTERING, UNIVERSAL_OPTIONS } from '@pnpm/common-cli-options-help'
 import { type Config, types as allTypes } from '@pnpm/config'
 import { PnpmError } from '@pnpm/error'
+import { type CheckDepsStatusOptions } from '@pnpm/deps.status'
 import {
   runLifecycleHook,
   makeNodeRequireOption,
@@ -184,6 +185,7 @@ export type RunOpts =
     }
     fallbackCommandUsed?: boolean
   }
+  & CheckDepsStatusOptions
 
 export async function handler (
   opts: RunOpts,
